backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
django-environ==0.11.2
django-extensions==3.2.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-storages==1.14.4

psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0

gunicorn==23.0.0
whitenoise==6.8.2

Pillow==11.1.0
numpy==2.2.2
scipy==1.15.1

drf-spectacular==0.28.0
django-import-export==4.3.4

python-dateutil==2.9.0
